Non- sequential cropping
This notebook helps identify potential areas of cropping by summing the rows and columns pixel values. Then identifying the width from the peaks of the sum of rows and columns. Using a minimum width from the histogram of diameter of the neurons measured earlier, we segregate the region of interests into horizontal and vertical regions to be be searched. Later these ROI are searched with 50 * 50 crops with the classifier.